VGG16 - FER2013
To use the VGG16 model for facial emotion recognition on the FER2013 dataset, which consists of grayscale images and emotion labels. After modifying the pre-trained VGG16 model to accommodate the seven emotion classes in FER2013, we can compile and train the model using an appropriate loss function and optimizer. Finally, evaluate the model's performance on a validation and test set to measure its accuracy in recognizing facial emotions.